Integrated the deploy trigger logic into the sdk package
