+++ cmd/console/deploy/deploy_test.go
@@ -7,14 +7,15 @@ import (
 	"net/url"
 	"testing"
 
+	"github.com/mia-platform/miactl/factory"
+
+	"github.com/mia-platform/miactl/sdk"
+
 	"github.com/mia-platform/miactl/renderer"
-	"github.com/mia-platform/miactl/sdk/deploy"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/h2non/gock.v1"
 )
-
-const triggeredPipelinesKey = "triggered-pipelines"
 
 func TestNewDeployCmd(t *testing.T) {
 	const (
@@ -26,9 +27,9 @@ func TestNewDeployCmd(t *testing.T) {
 	)
 	const expectedPipelineId = 458467
 	expectedPipelineURL := fmt.Sprintf("https://pipeline-url/%d", expectedPipelineId)
-	triggerEndpoint := fmt.Sprintf("/deploy/projects/%s/trigger/pipeline/", projectId)
+	triggerEndpoint := fmt.Sprintf("api/deploy/projects/%s/trigger/pipeline/", projectId)
 
-	expectedPipeline := deploy.Pipeline{
+	expectedPipeline := sdk.PipelineConfig{
 		ProjectId:   projectId,
 		PipelineId:  expectedPipelineId,
 		Environment: environment,
@@ -63,7 +64,8 @@ func TestNewDeployCmd(t *testing.T) {
 		cmd.Flags().Set("environment", environment)
 		cmd.Flags().Set("revision", revision)
 
-		err := cmd.ExecuteContext(context.Background())
+		ctx := factory.WithValue(context.Background(), cmd.OutOrStdout())
+		err := cmd.ExecuteContext(ctx)
 		require.NoError(t, err)
 
 		tableRows := renderer.CleanTableRows(buf.String())
@@ -74,10 +76,10 @@ func TestNewDeployCmd(t *testing.T) {
 		require.Equal(t, expectedHeaders, tableRows[0])
 		require.Equal(t, expectedRow, tableRows[1])
 
-		var triggeredPipelines deploy.Pipelines
+		var triggeredPipelines sdk.PipelinesConfig
 		require.NoError(t, viper.UnmarshalKey(triggeredPipelinesKey, &triggeredPipelines))
 		require.Equal(t, 1, len(triggeredPipelines), "Number of triggered pipelines should match")
-		require.Equal(t, expectedPipeline, triggeredPipelines[0], "Pipeline details should match")
+		require.Equal(t, expectedPipeline, triggeredPipelines[0], "PipelineConfig details should match")
 
 		require.True(t, gock.IsDone())
 	})
@@ -101,8 +103,8 @@ func TestNewDeployCmd(t *testing.T) {
 		viper.Set("apibaseurl", baseURL)
 		viper.Set("apitoken", apiToken)
 		viper.Set("project", projectId)
-		viper.Set(triggeredPipelinesKey, deploy.Pipelines{
-			deploy.Pipeline{ProjectId: "437t34b293u", PipelineId: 723531, Environment: "test"},
+		viper.Set(triggeredPipelinesKey, sdk.PipelinesConfig{
+			sdk.PipelineConfig{ProjectId: "437t34b293u", PipelineId: 723531, Environment: "test"},
 		})
 		viper.WriteConfigAs("/tmp/.miaplatformctl.yaml")
 
@@ -114,7 +116,8 @@ func TestNewDeployCmd(t *testing.T) {
 		cmd.Flags().Set("environment", environment)
 		cmd.Flags().Set("revision", revision)
 
-		err := cmd.ExecuteContext(context.Background())
+		ctx := factory.WithValue(context.Background(), cmd.OutOrStdout())
+		err := cmd.ExecuteContext(ctx)
 		require.NoError(t, err)
 
 		tableRows := renderer.CleanTableRows(buf.String())
@@ -125,7 +128,7 @@ func TestNewDeployCmd(t *testing.T) {
 		require.Equal(t, expectedHeaders, tableRows[0])
 		require.Equal(t, expectedRow, tableRows[1])
 
-		var triggeredPipelines deploy.Pipelines
+		var triggeredPipelines sdk.PipelinesConfig
 		require.NoError(t, viper.UnmarshalKey(triggeredPipelinesKey, &triggeredPipelines))
 		require.Equal(t, 2, len(triggeredPipelines), "Number of triggered pipelines should match")
 		require.Equal(t, expectedPipeline, triggeredPipelines[1], "Last pipeline details should match")
@@ -158,7 +161,8 @@ func TestNewDeployCmd(t *testing.T) {
 		cmd.Flags().Set("environment", environment)
 		cmd.Flags().Set("revision", revision)
 
-		err := cmd.ExecuteContext(context.Background())
+		ctx := factory.WithValue(context.Background(), cmd.OutOrStdout())
+		err := cmd.ExecuteContext(ctx)
 		require.NoError(t, err)
 
 		base, _ := url.Parse(baseURL)
@@ -169,7 +173,7 @@ func TestNewDeployCmd(t *testing.T) {
 			fmt.Sprintf("POST %s: 400", base.ResolveReference(path)),
 		)
 
-		var triggeredPipelines deploy.Pipelines
+		var triggeredPipelines sdk.PipelinesConfig
 		require.NoError(t, viper.UnmarshalKey(triggeredPipelinesKey, &triggeredPipelines))
 		require.Empty(t, triggeredPipelines, "No pipelines should be triggered")
 
@@ -195,10 +199,11 @@ func TestNewDeployCmd(t *testing.T) {
 		cmd.Flags().Set("environment", environment)
 		cmd.Flags().Set("revision", revision)
 
-		err := cmd.ExecuteContext(context.Background())
+		ctx := factory.WithValue(context.Background(), cmd.OutOrStdout())
+		err := cmd.ExecuteContext(ctx)
 		require.EqualError(t, err, "API base URL not specified nor configured")
 
-		var triggeredPipelines deploy.Pipelines
+		var triggeredPipelines sdk.PipelinesConfig
 		require.NoError(t, viper.UnmarshalKey(triggeredPipelinesKey, &triggeredPipelines))
 		require.Empty(t, triggeredPipelines, "No pipelines should be triggered")
 	})
@@ -221,10 +226,11 @@ func TestNewDeployCmd(t *testing.T) {
 		cmd.Flags().Set("environment", environment)
 		cmd.Flags().Set("revision", revision)
 
-		err := cmd.ExecuteContext(context.Background())
+		ctx := factory.WithValue(context.Background(), cmd.OutOrStdout())
+		err := cmd.ExecuteContext(ctx)
 		require.EqualError(t, err, "missing API token - please login")
 
-		var triggeredPipelines deploy.Pipelines
+		var triggeredPipelines sdk.PipelinesConfig
 		require.NoError(t, viper.UnmarshalKey(triggeredPipelinesKey, &triggeredPipelines))
 		require.Empty(t, triggeredPipelines, "No pipelines should be triggered")
 	})
